# optimized memcpy