This PR does a few things to improve memcpy performance:
  - branch for small copy lengths
  - when source and dest have the same address modulo suggested vector size, long copies are done via vectors suggested by `std.simd.suggestVectorLength(u8)`
  - when source and dest have different addresses modulo suggested vector size, the load is done from an aligned vector and the store to an `align(1)` vector
  
It is also possible to use aligned vector loads _and_ stores in the misaligned case, but I have not found a nice way to make the compiler do it (for x86_64), and would rather not do it with inline asm. The simplest way to do it (using `@shuffle`) does improve performance further, however it causes massive bloat in the memcpy function (producing >12x more code[^1] than without it when vectors have alignment 32) due to the need for a comptime mask.

Here are some throughput numbers showing performance improvements of 20-180% depending on the length copied. In summary, some lengths (in particular small powers of 2) have greatly improved performance, with long lengths (i.e. when not taking a branch for small length) having about 20% improvement. These numbers are from a micro benchmark that copies into a 32 byte aligned buffer (numbers are an average of copying starting at every offset 0..32 from a 32 byte aligned source buffer). The throughput numbers are for x86_64 with avx2 compiled in ReleaseFast for a handful of sizes; note that a branches for small sizes are taken for lengths < 256.

| copy length | master throughput | optimized throughput |
| -------------:     |    -----------------------:   |    --------------------------:     |
| 5 B | 1.08 GiB/s | 1.23 GiB/s |
| 8 B | 1.44 GiB/s | 4.08 GiB/s |
| 25 B | 3.65 GiB/s | 4.20 GiB/s |
| 32 B | 3.49 GiB/s | 6.05 GiB/s |
| 56 B | 7.65 GiB/s | 9.65 GiB/s |
| 64 B | 6.35 GiB/s | 12.5 GiB/s |
| 109 B | 10.65 GiB/s | 15.58 GiB/s |
| 128 B | 11.1 GiB/s | 22.26 GiB/s |
| 250 B | 16.87 GiB/s | 34.04 GiB/s |
| 256 B | 15.66 GiB/s | 19.54 GiB/s |
| 500 B | 19.82 GiB/s | 23.04 GiB/s |
| 512 B | 17.77 GiB/s | 23.33 GiB/s |
| 1 KiB | 19.39 GiB/s | 24.12 GiB/s |
| 4 KiB | 20.47 GiB/s | 24.28 GiB/s |

For code size of memcpy:
| optimize mode| master size (B) | optimized  size (B) |
| --------------------- |    ---------------:     |         ------------:          |
| Debug | 268 | 833 |
| ReleaseSafe | 251 | 830 |
| ReleaseFast | 251 | 830 |
| ReleaseSmall | 32 | 553 |

Much of the increased size comes from the code used for doing small copies (both for the total length being small and alignment fixups) - it could be tweaked if the speed/size tradeoff isn't considered worth it.

Note that I have only check the performance of this change on my (x86_64) machine - it would be good to check on other architectures as well if someone with one would like to test it.

[^1]: In ReleaseFast, using `@shuffle` to do aligned vector loads and stores in the misaligned case made `memcpy` 10204 bytes.