@Rexicon226 2024-02-13
Nice work! Could you share the benchmark script? I am really curious to see how those aligned loads/stores are emitted.