@dweiller 2024-02-13
> Nice work! Could you share the benchmark script? I am really curious to see how those aligned loads/stores are emitted.

The benchmark script is just a pretty simple loop copying from one buffer to the other a bunch of times for each source alignment offset. I haven't done anything clever in the micro benchmark. The reason aligned vector moves are emitted is the use of `@ptrCast` to get pointers of type `[*]CopyType`/`[*]const CopyType` in `lib/compiler_rt/memcpy.zig`. 

<details>
<summary> Here's the micro benchmark (doesn't look like I can colour it inside the fold):</summary>
<pre>
pub fn main() !void {
    const allocator = std.heap.page_allocator;

    const args = try std.process.argsAlloc(allocator);
    defer allocator.free(args);
    if (args.len > 3 or args.len < 2) return error.BadArgs;

    const iterations = try std.fmt.parseInt(usize, args[1], 10);

    const copy_len = if (args.len >= 3)
        try std.fmt.parseInt(usize, args[2], 10)
    else
        std.mem.page_size;

    std.debug.print("copying blocks of size {d}\n", .{std.fmt.fmtIntSizeBin(copy_len)});

    const alignment = if (std.simd.suggestVectorLength(u8)) |vec_size|
        @alignOf(@Type(.{ .Vector = .{ .child = u8, .len = vec_size } }))
    else
        @alignOf(usize);

    var times: [alignment]u64 = undefined;

    // preheat
    try iteration(allocator, alignment, copy_len, alignment - 2, &times[0], iterations);

    for (0..alignment) |s_offset| {
        try iteration(allocator, alignment, copy_len, s_offset, &times[s_offset], iterations);
        std.debug.print(
            "s_offset: {d}, time: {}\n",
            .{
                s_offset,
                std.fmt.fmtDuration(times[s_offset]),
            },
        );
    }

    var avg: u128 = 0;
    for (times) |t| {
        avg += t;
    }
    avg /= times.len;

    const final: u64 = @intCast(avg);

    const throughput: u64 = @intCast((@as(u128, copy_len) * iterations * std.time.ns_per_s) / final);

    std.debug.print("average time: {}, throughput: {d:.2}\n", .{
        std.fmt.fmtDuration(final),
        std.fmt.fmtIntSizeBin(throughput),
    });
}

inline fn iteration(
    allocator: Allocator,
    comptime alignment: comptime_int,
    copy_len: usize,
    s_offset: usize,
    p_time: *u64,
    iterations: usize,
) !void {
    const src = try allocator.alignedAlloc(u8, alignment, copy_len + s_offset);
    defer allocator.free(src);

    const dest = try allocator.alignedAlloc(u8, alignment, copy_len);
    defer allocator.free(dest);

    var timer = try std.time.Timer.start();

    var i: usize = 0;
    while (i < iterations) : (i += 1) {
        @memcpy(dest[0..copy_len], src[s_offset..][0..copy_len]);
    }

    const time = timer.read();

    p_time.* = time;
}

const std = @import("std");
const Allocator = std.mem.Allocator;
</pre>
</details>