@dweiller 2024-02-13
> I recommend you use something like:
> 
> ```zig
> dest[0..v].* = @as(*const @Vector(v, u8), @alignCast(@ptrCast(aligned_src[offset..][0..v]))).*;
> ```
> 
> to generate your aligned vector loads, 😉

I assume you mean `@as(*Vector(V, u8), @alignCast(@ptrCast(dest[0..v]))).* = ...`? Otherwise that is going to produce an unaligned store. Using this particular snippet to perform the case (casting the dest as well) doesn't make a difference.