@Jarred-Sumner 2024-02-13
I suspect zig’s memcpy could get a whole lot faster than 30%-ish. I think it’s hard to rule out measurement noise without also having a complete benchmark and on larger amounts of data. There must already be comprehensive memcpy benchmarks online somewhere you could copy from