@dweiller 2024-02-14
> I suspect zig’s memcpy could get a whole lot faster than 30%-ish.

For small sizes this is certainly true, but for large copies I'm a bit skeptical, at least not without using inline asm with something like `rep movsb`, but that opens up a bit of a can of worms.