@dweiller 2024-02-14
I've made a bunch of improvements yielding both better performance and smaller code size. The table in the top post has been updated.